+++ src/gui/interface.py
@@ -18,6 +18,7 @@ logging.basicConfig(
 
 BAUDRATE = 115_200
 DEVICE_PORT: int = int()
+PORT: str = str()
 
 
 def find_device(DEVICE_NUMBER: int = DEVICE_PORT) -> tuple[str, list[str]]:
@@ -30,7 +31,7 @@ def find_device(DEVICE_NUMBER: int = DEVICE_PORT) -> tuple[str, list[str]]:
         find = ["find /dev -iname '*acm*'"]
         try:
             logger.info(
-                msg=f"{__name__} executing linux shell command to get device file names"
+                msg=f"{find_device.__name__} function executing linux shell command to get device file names"
             )
             results_ = subprocess.run(
                 args=find,
@@ -43,12 +44,15 @@ def find_device(DEVICE_NUMBER: int = DEVICE_PORT) -> tuple[str, list[str]]:
             results = sorted(results_.stdout.strip().splitlines())
             global PORT
             PORT = results[DEVICE_NUMBER]
+            logger.info(msg=f"Connected Devices: {results}")
+            logger.debug(msg=f"Chosen Device: {PORT}")
             return results[DEVICE_NUMBER], results
         except IndexError:
             logger.warning("No device connected to machine")
             logger.exception("Device not found")
 
     elif platform.system().lower() == "windows":
+        logger.info(f"{platform.system()} discovered")
         # global PORT
         # Search Windows filesystem for device
         # filename = "COM3"
@@ -58,17 +62,23 @@ def find_device(DEVICE_NUMBER: int = DEVICE_PORT) -> tuple[str, list[str]]:
         return "COM3"
 
 
-def serial_call(*args, PORT=DEVICE_PORT) -> None:
+def serial_call(*args) -> None:
     logger.debug(msg="Serial call to device initiated")
-    logger.info(msg=f"Serial call initiated from {platform.system()} to USB device")
+    # logger.info(msg=f"Serial call initiated from {platform.system()} to {find_device()[DEVICE_PORT]}")
     sleep(0.1)  # Allow time to read and execute via serial
+    global PORT
     try:
-        with serial.Serial() as ser:
+        with serial.Serial() as ser:  #  If the teensy is corrupted the GUI freezes here
             ser.baudrate = BAUDRATE
+            logger.debug(msg=f"baudrate set from {serial_call.__name__}")
             ser.port = PORT
+            logger.debug(msg=f"Serial device set from {serial_call.__name__}")
             ser.timeout = 2  # seconds
+            logger.debug(msg=f"timeout set from {serial_call.__name__}")
             ser.open()
+            logger.debug(msg=f"opening serial connection from {serial_call.__name__}")
             ser.write(f"{' '.join([arg for arg in args])}".encode("utf-8"))
+            logger.debug(msg=f"information send via serial protocol; {serial_call.__name__}")
             # outputs: list[str] = [line.decode('ascii') for line in ser.readlines()] # Serial output from MGTron
             # outputs = " ".join(output for output in outputs)
             # print(outputs)
@@ -80,11 +90,13 @@ def serial_call(*args, PORT=DEVICE_PORT) -> None:
 class Megatron:
     """Class to organize the manipulation of 8 channels"""
 
+    logger.info(msg="\n\nGUI LAUNCHED\n\n")
+
     try:
         logger.info(msg="Getting the port name of device")
         global PORT
         PORT = find_device(DEVICE_PORT)[0]
-        print("PORT: ", PORT)
+        logger.info(msg=f"Connected device path: {PORT}")
     except TypeError:
         logger.error(msg="No device found on system")
         logger.exception(msg="No device found on system")
@@ -92,8 +104,8 @@ class Megatron:
     def status(self) -> None:
         """Check the status of the board"""
 
-        serial_call("s", PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("s")
+        logger.info(f"{Megatron.status.__name__} function executed")
 
     def change_power(self, channel: int, power_level: int) -> None:
         """
@@ -101,8 +113,9 @@ class Megatron:
         Range: 0 - 63
         """
 
-        serial_call("p", str(channel), str(power_level), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("p", str(channel), str(power_level))
+        logger.info(f"{Megatron.change_power.__name__} function executed")
+        logger.info(msg=f"Connected device path: {PORT}")
 
     def change_freq(self, channel: int, frequency: float) -> None:
         """
@@ -110,8 +123,8 @@ class Megatron:
         Range: 50 - 6400 MHz
         """
 
-        serial_call("f", str(channel), str(frequency), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("f", str(channel), str(frequency))
+        logger.info(f"{Megatron.change_freq.__name__} function executed")
 
     def change_bandwidth(self, channel: int, percentage: int) -> None:
         """
@@ -119,29 +132,32 @@ class Megatron:
         Range: 0 - 100
         """
 
-        serial_call("b", str(channel), str(percentage), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("b", str(channel), str(percentage))
+        logger.info(f"{Megatron.change_bandwidth.__name__} function executed")
 
     def save_state(self, state: bool) -> None:
         """Save each settings made by the user into memory for next startup"""
 
         state = 1 if state else 0
-        serial_call("x", str(state), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        try:
+            serial_call("x", str(state))
+            logger.info(f"{Megatron.save_state.__name__} function executed")
+        except TypeError:
+            logger.exception(msg="No device assigned")
 
     def amplification(self, channel: int, state: bool) -> None:
         """Output HIGH or LOW logic level out of a chosen channel"""
 
         state = 1 if state else 0
-        serial_call("a", str(channel), str(state), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("a", str(channel), str(state))
+        logger.info(f"{Megatron.amplification.__name__} function executed")
 
     def stability(self, state: bool) -> None:
         """Boolean a second filtering stage of capacitors for further stability"""
 
         state = 1 if state else 0
-        serial_call("~", str(state), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("~", str(state))
+        logger.info(f"{Megatron.stability.__name__} function executed")
 
     def noise_control(self, state: bool, percentage: int) -> None:
         """
@@ -151,22 +167,27 @@ class Megatron:
         """
 
         state = 1 if state else 0
-        serial_call("n", str(state), str(percentage), PORT=PORT)
-        logger.info(f"{__name__} executed")
+        serial_call("n", str(state), str(percentage))
+        logger.info(f"{Megatron.noise_control.__name__} function executed")
 
     def reset_board(self) -> None:
         """Reset the parameters of the board"""
 
         [
             (
-                serial_call("p", str(i), "0", PORT=PORT),
-                serial_call("b", str(i), "0", PORT=PORT),
-                serial_call("f", str(i), "50.00", PORT=PORT),
+                serial_call("p", str(i), "0"),
+                serial_call("b", str(i), "0"),
+                serial_call("f", str(i), "50.00"),
             )
             for i in range(1, 9)
         ]
 
-    logger.info(f"{__name__} executed")
+        logger.info(f"{Megatron.reset_board.__name__} function executed")
+
+    logger.info(msg="class Megatron initialized")
+
+
+logger.debug(msg="EOF")
 
 
 def main() -> None:
@@ -174,7 +195,6 @@ def main() -> None:
 
     # find_device("linux")
     # test_1 = Megatron()
-    print(platform.system().lower())
 
     # for i in range(8):
     # test_1.change_power(i+1, random.randint(a=10, b=63))